test-requirements.txt
hacking<0.11,>=0.10.0
pytest
pytest-cov
pytest-mock<0.10.0
pyzabbix==0.7.3
